+++ wcag-tool/backend/src/models/annotation.model.js
@@ -5,11 +5,11 @@ const annotationSchema = new Schema(
   {
     title: {
       type: String,
-      required: true,
+      required: false,
     },
     description: {
       type: String,
-      required: true,
+      required: false,
     },
     selector: {
       type: String,
