+++ wcag-tool/chrome-extension/public/editor.html
@@ -6,7 +6,7 @@
     <script src="js/vendor.js"></script>
   </head>
   <body>
-    <div id="root" class="h-screen w-96"></div>
+    <div id="root" class="h-screen"></div>
     <iframe
       id="snapshot-iframe"
       class="editor"
